kernel/settings/deployment.py
from .base import *
from .secure import *
from .packages import *

from decouple import config

DEBUG = config('DEBUG', cast=bool)
ALLOWED_HOSTS = config('ALLOWED_HOSTS', cast=lambda v: [s.strip() for s in v.split(',')])
